refactor(winner): 접근제한자, 형변환 위치 변경

+++ src/main/java/lotto/Lotto.java
@@ -15,8 +15,7 @@ public class Lotto {
 	}
 
 	public Winner winningCheck(Lotto winLotto) {
-		return Winner.valueOf(
-			(int) numbers.stream()
+		return Winner.valueOf(numbers.stream()
 				.filter(winLotto.numbers::contains)
 				.count());
 	}
